# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Rapid Force Cyber Fusion AI Factory API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type AgentRole = (typeof AgentRole)[keyof typeof AgentRole];

export const AgentRole = {
  architect: "architect",
  cto: "cto",
  product_manager: "product_manager",
  security_lead: "security_lead",
  senior_engineer: "senior_engineer",
  devops: "devops",
  ui_ux: "ui_ux",
  red_team: "red_team",
  documentation: "documentation",
} as const;

export type AgentStatus = (typeof AgentStatus)[keyof typeof AgentStatus];

export const AgentStatus = {
  active: "active",
  idle: "idle",
  standby: "standby",
  offline: "offline",
} as const;

export interface Agent {
  id: number;
  name: string;
  role: AgentRole;
  module: string;
  status: AgentStatus;
  missionsCompleted: number;
  tasksActive: number;
  lastHeartbeat: string;
  createdAt: string;
}

export type CreateAgentBodyRole =
  (typeof CreateAgentBodyRole)[keyof typeof CreateAgentBodyRole];

export const CreateAgentBodyRole = {
  architect: "architect",
  cto: "cto",
  product_manager: "product_manager",
  security_lead: "security_lead",
  senior_engineer: "senior_engineer",
  devops: "devops",
  ui_ux: "ui_ux",
  red_team: "red_team",
  documentation: "documentation",
} as const;

export interface CreateAgentBody {
  name: string;
  role: CreateAgentBodyRole;
  module: string;
}

export type UpdateAgentBodyStatus =
  (typeof UpdateAgentBodyStatus)[keyof typeof UpdateAgentBodyStatus];

export const UpdateAgentBodyStatus = {
  active: "active",
  idle: "idle",
  standby: "standby",
  offline: "offline",
} as const;

export interface UpdateAgentBody {
  status?: UpdateAgentBodyStatus;
  tasksActive?: number;
}

export interface AgentsSummary {
  total: number;
  active: number;
  idle: number;
  standby: number;
  offline: number;
  totalMissionsCompleted: number;
}

export type SprintStatus = (typeof SprintStatus)[keyof typeof SprintStatus];

export const SprintStatus = {
  planning: "planning",
  active: "active",
  complete: "complete",
  paused: "paused",
} as const;

export interface Sprint {
  id: number;
  name: string;
  objective: string;
  status: SprintStatus;
  startDate: string;
  endDate: string;
  progress: number;
  createdAt: string;
}

export interface CreateSprintBody {
  name: string;
  objective: string;
  startDate: string;
  endDate: string;
}

export type UpdateSprintBodyStatus =
  (typeof UpdateSprintBodyStatus)[keyof typeof UpdateSprintBodyStatus];

export const UpdateSprintBodyStatus = {
  planning: "planning",
  active: "active",
  complete: "complete",
  paused: "paused",
} as const;

export interface UpdateSprintBody {
  status?: UpdateSprintBodyStatus;
  progress?: number;
}

export type MissionStatus = (typeof MissionStatus)[keyof typeof MissionStatus];

export const MissionStatus = {
  active: "active",
  pending: "pending",
  complete: "complete",
  failed: "failed",
} as const;

export type MissionPriority =
  (typeof MissionPriority)[keyof typeof MissionPriority];

export const MissionPriority = {
  critical: "critical",
  high: "high",
  medium: "medium",
  low: "low",
} as const;

export interface Mission {
  id: number;
  title: string;
  description: string;
  status: MissionStatus;
  priority: MissionPriority;
  assignedAgentId?: number | null;
  sprintId?: number | null;
  category: string;
  createdAt: string;
  updatedAt: string;
}

export type CreateMissionBodyPriority =
  (typeof CreateMissionBodyPriority)[keyof typeof CreateMissionBodyPriority];

export const CreateMissionBodyPriority = {
  critical: "critical",
  high: "high",
  medium: "medium",
  low: "low",
} as const;

export interface CreateMissionBody {
  title: string;
  description: string;
  priority: CreateMissionBodyPriority;
  assignedAgentId?: number;
  sprintId?: number;
  category: string;
}

export type UpdateMissionBodyStatus =
  (typeof UpdateMissionBodyStatus)[keyof typeof UpdateMissionBodyStatus];

export const UpdateMissionBodyStatus = {
  active: "active",
  pending: "pending",
  complete: "complete",
  failed: "failed",
} as const;

export interface UpdateMissionBody {
  status?: UpdateMissionBodyStatus;
  assignedAgentId?: number;
}

export type ThreatSeverity =
  (typeof ThreatSeverity)[keyof typeof ThreatSeverity];

export const ThreatSeverity = {
  critical: "critical",
  high: "high",
  medium: "medium",
  low: "low",
} as const;

export type ThreatStatus = (typeof ThreatStatus)[keyof typeof ThreatStatus];

export const ThreatStatus = {
  active: "active",
  mitigated: "mitigated",
  monitoring: "monitoring",
  closed: "closed",
} as const;

export interface Threat {
  id: number;
  title: string;
  description: string;
  severity: ThreatSeverity;
  category: string;
  source: string;
  status: ThreatStatus;
  detectedAt: string;
  createdAt: string;
}

export type CreateThreatBodySeverity =
  (typeof CreateThreatBodySeverity)[keyof typeof CreateThreatBodySeverity];

export const CreateThreatBodySeverity = {
  critical: "critical",
  high: "high",
  medium: "medium",
  low: "low",
} as const;

export interface CreateThreatBody {
  title: string;
  description: string;
  severity: CreateThreatBodySeverity;
  category: string;
  source: string;
}

export interface ThreatsSummary {
  total: number;
  critical: number;
  high: number;
  medium: number;
  low: number;
  mitigated: number;
  active: number;
}

export interface ActivityEvent {
  id: number;
  type: string;
  message: string;
  agentName?: string | null;
  entityType?: string | null;
  entityId?: number | null;
  createdAt: string;
}

export type ListMissionsParams = {
  status?: ListMissionsStatus;
  sprintId?: number;
};

export type ListMissionsStatus =
  (typeof ListMissionsStatus)[keyof typeof ListMissionsStatus];

export const ListMissionsStatus = {
  active: "active",
  pending: "pending",
  complete: "complete",
  failed: "failed",
} as const;

export type ListThreatsParams = {
  severity?: ListThreatsSeverity;
};

export type ListThreatsSeverity =
  (typeof ListThreatsSeverity)[keyof typeof ListThreatsSeverity];

export const ListThreatsSeverity = {
  critical: "critical",
  high: "high",
  medium: "medium",
  low: "low",
} as const;

export type ListActivityParams = {
  limit?: number;
};